requirements.txt
AfricastalkingGateway==1.8
appdirs==1.4.0
Babel==2.3.4
BabelDjango==0.2.2
boto3==1.4.4
botocore==1.5.4
braintree==3.34.0
cffi==1.9.1
colorama==0.3.9
cryptography==1.7.1
cssselect==1.0.1
cssutils==1.0.1
dj-database-url==0.4.2
dj-email-url==0.0.10
Django==1.10.7
django-absolute==0.3
django-bootstrap3==8.2.1
django-cache-url==1.3.1
django-compressor==2.2
django-countries==4.1
django-emailit==0.2.2
django-materializecss-form==1.0.1
django-model-utils==3.0.0
django-mptt==0.8.7
django-notifications-hq==1.3
django-payments==0.11.0.3
django-prices==0.4.13
django-prices-openexchangerates==0.1.12
django-redis==4.7.0
django-storages==1.5.2
django-versatileimagefield==1.6.2
django-webpack-loader==0.4.1
django-xhtml2pdf==0.0.3
djangorestframework==3.5.3
djangorestframework-jwt==1.10.0
docutils==0.13.1
elasticsearch==2.4.1
enum34==1.1.6
Faker==0.7.7
funcsigs==1.0.2
future==0.16.0
futures==3.1.1
google-i18n-address==2.0.1
google-measurement-protocol==0.1.6
graphene==2.0.1
graphene-django==2.0.0
graphql-core==2.0
graphql-relay==0.4.5
html5lib==1.0b8
idna==2.2
ipaddress==1.0.18
iso8601==0.1.11
jmespath==0.9.0
jsonfield==1.0.3
lxml==3.8.0
Markdown==2.6.7
maxminddb==1.2.3
maxminddb-geolite2==2015.708
mock==2.0.0
oauthlib==2.0.1
packaging==16.8
pbr==1.10.0
Pillow==3.4.2
premailer==3.0.1
prices==0.5.8
promise==2.1 # via graphene, graphene-django, graphql-core, graphql-relay
psycopg2==2.7.5
psycopg2-binary==2.7.5
purl==1.3
py==1.4.32
pyasn1==0.1.9
pycparser==2.17
PyInstaller==3.2.1
PyJWT==1.4.2
pyparsing==2.1.10
PyPDF2==1.26.0
#pypiwin32==219
pytest==3.0.5
pytest-django==3.1.2
python-dateutil==2.6.0
python-openid==2.2.5
pytz==2016.10
redis==2.10.5
reportlab==3.4.0
requests==2.12.4
requests-oauthlib==0.7.0
s3transfer==0.1.10
satchless==1.1.3
simplejson==3.10.0
singledispatch==3.4.0.3
six==1.10.0
social-auth-app-django==1.1.0
social-auth-core==1.2.0
stripe==1.46.0
structlog=18.2.0
suds-jurko==0.6
text-unidecode==1.0
typing==3.5.2.2
urllib3==1.19.1
xhtml2pdf==0.0.6
xmltodict==0.10.2
django-macaddress==1.5.0
pycrypto==2.6.1
netifaces==0.10.6
